% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc.basedata.R
\name{basedata.year}
\alias{basedata.year}
\title{basedata.year}
\usage{
basedata.year(ayear, abasedata.allyear)
}
\arguments{
\item{year}{data of a numeric year}

\item{basedata.allyear}{results from dataproc.basedata()}
}
\value{
A list of base data in a year: basedata.regmkt, basedata.trade, basedata.pricelink
}
\description{
basedata.year
}
