% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc.basedata.R
\name{basedata_year}
\alias{basedata_year}
\title{basedata_year}
\usage{
basedata_year(YEAR, BASEDATA.ALLYEARS)
}
\arguments{
\item{YEAR}{data of a numeric year}

\item{BASEDATA.ALLYEARS}{results from dataproc_basedata()}
}
\value{
A list of base data in a year: basedata.regmkt, basedata.trade, basedata.pricelink
}
\description{
basedata_year
}
