% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.logit.CES.R
\name{ces.share}
\alias{ces.share}
\title{ces.share}
\usage{
ces.share(p, ces.exponent, share.weight)
}
\arguments{
\item{p}{A vector of prices in the nest}

\item{ces.exponent}{A numeric value of CES exponent}

\item{share.weight}{Share-weights to be used in the function}
}
\value{
A vector of value shares from the logit sharing function
}
\description{
A constant elasticity of substitution function generating value sharing
}
