% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc.basedata.R
\name{dataproc.basedata}
\alias{dataproc.basedata}
\title{BASEDATA}
\usage{
dataproc.basedata()
}
\value{
a list of basedata.regmkt.allyear, basedata.trade.allyear, and basedata.pricelink.allyear
}
\description{
processes source data to generate basedata
}
\author{
Xin Zhao 2021
}
