% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc.gtap.R
\name{dataproc.gtap}
\alias{dataproc.gtap}
\title{dataproc.gtap}
\usage{
dataproc.gtap(
  gtapdatadir = paste0(system.file("extdata", package = "tradecast", mustWork = T),
    "/gtap/")
)
}
\arguments{
\item{datadir}{path to data folder (default = "./inst/extdata/gtap/")}
}
\value{
Returns a dataframe (basedata.pricelink.margin.mtax) including margin and
        tariff, and also nonland cost  share.
}
\description{
processes gtap raw data to get margin and tariff data for tradecast
}
\author{
Xin Zhao 2021
}
