% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.script.R
\name{dslnex}
\alias{dslnex}
\title{Partial equilibrium agricultural economic modeling (dslnex)}
\usage{
dslnex(x)
}
\arguments{
\item{x}{price variables}
}
\value{
The slack variables (y) in system of equations
}
\description{
The model includes system of equations representing global agricultural markets.
             There are currently 144 variable and 144 equations to solve.
}
