% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.script.R
\name{dslnex_result}
\alias{dslnex_result}
\title{dslnex_result}
\usage{
dslnex_result(x = model.sol$x, output = T)
}
\arguments{
\item{x}{price variables}

\item{output}{A logical variable: with default "T", other results are output}
}
\value{
A list of output results
}
\description{
Using the same code in dslnex to process solved prices and get other results
}
