% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.logit.CES.R
\name{logit.share}
\alias{logit.share}
\title{logit.share}
\usage{
logit.share(p, logit.exponent, share.weight)
}
\arguments{
\item{p}{A vector of prices in the nest}

\item{logit.exponent}{A numeric value of logit exponent}

\item{share.weight}{Share-weights to be used in the function}
}
\value{
A vector of volume shares from the logit sharing function
}
\description{
A function generating relative cost logit sharing given parameters and prices
}
