% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_script_minicam_agtrade.R
\name{minicam_agtrade}
\alias{minicam_agtrade}
\title{Partial equilibrium agricultural economic modeling (minicam_agtrade)}
\usage{
minicam_agtrade(X, CONFIG, OUTPUT = F)
}
\arguments{
\item{X}{price variables}

\item{CONFIG}{The configuration generated from model.shock.config}

\item{OUTPUT}{A logical variable: with default "T", other results are output}
}
\value{
The slack variables (y) in system of equations
}
\description{
The model includes system of equations representing global agricultural markets.
             There are currently 144 variable and 144 equations to solve.
}
