% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.proc.R
\name{model.proc.output}
\alias{model.proc.output}
\title{model.proc.output}
\usage{
model.proc.output(
  base.yr = 1995,
  target.yr = 1995,
  aHISTDATA,
  output,
  abasedata.allyear
)
}
\arguments{
\item{base.yr}{A numeric value of base year}

\item{target.yr}{A numeric value of target year}

\item{aHISTDATA}{list of all data, returned from HISTDATA()}

\item{output}{updated.db.equil.yr}

\item{abasedata.allyear}{basedata.allyear returned from dataproc.basedata}
}
\value{

}
\description{
model.proc.output
}
