% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.config.R
\name{model.shock.config}
\alias{model.shock.config}
\title{model.shock.config}
\usage{
model.shock.config(base.yr = 1995, target.yr = 2015, aHISTDATA)
}
\arguments{
\item{base.yr}{A numeric value of base year}

\item{target.yr}{A numeric value of target year to calculate shocks or targeted exogenous variables}

\item{aHISTDATA}{list of all data, returned from HISTDATA()}
}
\value{
A list of exogenous variable and parameters to be used in model solving
}
\description{
generate a list model.shock.config of model drivers in global env.
}
