% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.solve.R
\name{model.solve}
\alias{model.solve}
\title{model.solve}
\usage{
model.solve(
  config,
  xstart = rep(100, 144),
  abtol = 0.01,
  amaxit = 1000,
  amessage = T
)
}
\arguments{
\item{config}{config generated from model.shock.config}

\item{xstart}{A numeric vector with initial values for solving, e.g., rep(100,144)}

\item{abtol}{Passed to btol in nleqslv}

\item{amaxit}{Passed to maxit in nleqslv}

\item{amessage}{If T, print solution message}
}
\value{
Model solution
}
\description{
Solver of the model, wrapped from nleqslv.
}
