% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data.R
\name{model_data}
\alias{model_data}
\title{model.data}
\usage{
model_data(
  HIST.YEARS = CALIBRATION_YEAR,
  BASEDATA.ALLYEARS,
  PARAMETER.EXPONENT
)
}
\arguments{
\item{HIST.YEARS}{A numeric value of a historical year; CALIBRATION_YEAR is used by default}

\item{BASEDATA.ALLYEARS}{Results from dataproc.basedata()}

\item{PARAMETER.EXPONENT}{A list of parameters used in initial calibration (e.g., CES and logit exponents)}
}
\value{
A list of data including parameters and data across HIST.YEARS
}
\description{
Making use of the processed data to balance base year equilibrium and return essential data needed for initializing and verifying the model.
}
