% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_hindcast_S2.R
\name{model_hindcast_S2}
\alias{model_hindcast_S2}
\title{model_hindcast_S2}
\usage{
model_hindcast_S2(
  PARAMETER,
  PARAMETER.EXPONENT,
  BASEYEAR,
  TARGETYEARS,
  BASEDATA.ALLYEARS,
  LOG.WEIGHT = F
)
}
\arguments{
\item{PARAMETER}{A vector of parameters to optimize}

\item{BASEYEAR}{A numeric value of base year}

\item{TARGETYEARS}{A vector of target years}

\item{BASEDATA.ALLYEARS}{The dataframe returned from dataproc.basedata}

\item{LOG.WEIGHT}{if true using log(1+weight) as weight; otherwise weight^0.5 is used as weight in output_metric}
}
\value{
goodness-of-fit (GoF)
}
\description{
Generate the FUN (returning GoF) to be used in hindcast optimization for S2
}
