% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_hindcast_S3.R
\name{model_hindcast_S3}
\alias{model_hindcast_S3}
\title{model_hindcast_S3}
\usage{
model_hindcast_S3(
  PARAMETER,
  PARAMETER.EXPONENT,
  BASEYEAR,
  TARGETYEARS,
  BASEDATA.ALLYEARS
)
}
\arguments{
\item{PARAMETER}{A vector of parameters to optimize}

\item{BASEYEAR}{A numeric value of base year}

\item{TARGETYEARS}{A vector of target years}

\item{BASEDATA.ALLYEARS}{The dataframe returned from dataproc.basedata}
}
\value{
goodness-of-fit (GoF)
}
\description{
Generate the FUN (returning GoF) to be used in hindcast optimization for S3
}
