% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_proc_output.R
\name{model_proc_output}
\alias{model_proc_output}
\title{model.proc.output}
\usage{
model_proc_output(
  BASEYEAR,
  TARGETYEAR,
  MODEL.DATA,
  DB.OUTPUT,
  BASEDATA.ALLYEARS
)
}
\arguments{
\item{BASEYEAR}{A numeric value of base year}

\item{TARGETYEAR}{A numeric value of target year}

\item{MODEL.DATA}{list of all data, returned from MODEL.DATA()}

\item{DB.OUTPUT}{Output database from model_solve}

\item{BASEDATA.ALLYEARS}{BASEDATA.ALLYEARS returned from dataproc.basedata}
}
\value{
Processed output database for a target year (updated.db.equil.yr)
}
\description{
model.proc.output
}
