% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_shock_S2.R
\name{model_shock_S2}
\alias{model_shock_S2}
\title{model_shock_S2}
\usage{
model_shock_S2(BASEYEAR = 1995, TARGETYEAR = 2015, MODEL.DATA)
}
\arguments{
\item{BASEYEAR}{A numeric value of base year}

\item{TARGETYEAR}{A numeric value of target year to calculate shocks or targeted exogenous variables}

\item{MODEL.DATA}{list of all data, returned from MODEL.DATA()}
}
\value{
A list of exogenous variable and parameters to be used in model solving
}
\description{
Translate MODEL.DATA to model language of configuration info,
             e.g., parameters and shocks. Sourcing fn.demand.regl.sw from target
}
