% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_solve.R
\name{model_solve}
\alias{model_solve}
\title{model.solve}
\usage{
model_solve(
  CONFIG,
  FUN,
  XSTART = rep(100, 144),
  BTOL = 0.01,
  MAXIT = 1000,
  MESSAGE = T
)
}
\arguments{
\item{CONFIG}{The configuration generated from model.shock.config}

\item{FUN}{A function (passed to nleqslv) of XSTART returning a vector of function values with the same length as XSTART}

\item{XSTART}{A numeric vector with initial values for solving, e.g., rep(100,144)}

\item{BTOL}{Passed to btol in nleqslv}

\item{MAXIT}{Passed to maxit in nleqslv}

\item{MESSAGE}{If T, print solution message}
}
\value{
A list of model solution (model.sol) and output data (outDB)
}
\description{
Solver of the model, wrapped from nleqslv.
}
