% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.metric.R
\name{output.metric}
\alias{output.metric}
\title{output.metric}
\usage{
output.metric(aupdated.db.equil)
}
\arguments{
\item{aupdated.db.equil}{processed output database from output.prod (updated.db.equil)}
}
\value{

}
\description{
output.metric
}
