% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_metric.R
\name{output_metric}
\alias{output_metric}
\title{output_metric}
\usage{
output_metric(DF)
}
\arguments{
\item{DF}{processed output database from hindcast experiment}
}
\value{
The value of targeted error metric
}
\description{
Computes goodness-of-fit based on the output database
}
