% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.dataproc.R
\name{pull_list}
\alias{pull_list}
\title{pull_list}
\usage{
pull_list(.df, group = "reg", val.rm = c("variable", "crop"))
}
\arguments{
\item{.df}{data frame}

\item{group}{grouping variable to pull as vectors}

\item{val.rm}{variables will be removed}
}
\value{
list of vectors
}
\description{
converting data frame to list for modeling purposes
}
