% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_hindcast.R
\name{run_hindcast}
\alias{run_hindcast}
\title{run_hindcast}
\usage{
run_hindcast(
  BASEYEAR,
  TARGETYEARS,
  MODEL.DATA,
  BASEDATA.ALLYEARS,
  MODEL = minicam_agtrade
)
}
\arguments{
\item{BASEYEAR}{A numeric value of base year}

\item{TARGETYEARS}{A vector of target years}

\item{MODEL.DATA}{list of all data, returned from MODEL.DATA()}

\item{BASEDATA.ALLYEARS}{The dataframe returned from dataproc.basedata}

\item{MODEL}{A function including model equations}
}
\value{
output database and print running time
}
\description{
run_hindcast
}
