% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_hindcast_optim.R
\name{run_hindcast_optim}
\alias{run_hindcast_optim}
\title{run_hindcast_optim}
\usage{
run_hindcast_optim(
  PAR.START = c(1.46, 1.5),
  FUN = model_hindcast,
  ...,
  LOWER = c(0.1, 0.1),
  UPPER = c(5, 10),
  HESSIAN = F,
  PGTOL = 1e-04,
  MAXIT = 1000
)
}
\arguments{
\item{PAR.START}{initial parameter in optim}

\item{FUN}{fn function in optim}

\item{...}{Pass through varibles to be used in FUN}

\item{LOWER}{Lower bound of parameter}

\item{UPPER}{Uppler bound of parameter}

\item{HESSIAN}{If T, producing Hessian metrix}

\item{PGTOL}{Tolerance in optim for the "L-BFGS-B" method}

\item{MAXIT}{Number of max iteration in optim}
}
\value{
Solutions from the optimization
}
\description{
A wrapper of the optim function to be used for hindcast optimization
}
