% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HISTDATA.R
\name{HISTDATA}
\alias{HISTDATA}
\title{HISTDATA}
\usage{
HISTDATA(hist.years = CALIBRATION_YEAR, abasedata.allyear)
}
\arguments{
\item{hist.years}{A numeric value of a historical year; CALIBRATION_YEAR is used by default}

\item{abasedata.allyear}{results from dataproc.basedata()}
}
\value{
A list of data including parameters and data across hist.years
}
\description{
Making use of the processed data to balance base year equilibrium and return essential data needed for initializing and verifying the model.
}
