% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_hindcast.R
\name{model_hindcast}
\alias{model_hindcast}
\title{model_hindcast}
\usage{
model_hindcast(
  PARAMETER,
  PARAMETER.EXPONENT,
  BASEYEAR,
  TARGETYEARS,
  BASEDATA.ALLYEARS,
  LOG.WEIGHT = F,
  DB.OUTPUT = F
)
}
\arguments{
\item{PARAMETER}{A vector of parameters to optimize}

\item{PARAMETER.EXPONENT}{A list of parameters used in initial calibration (e.g., CES and logit exponents)}

\item{BASEYEAR}{A numeric value of base year}

\item{TARGETYEARS}{A vector of target years}

\item{BASEDATA.ALLYEARS}{The dataframe returned from dataproc.basedata}

\item{LOG.WEIGHT}{if true using log(1+weight) as weight; otherwise weight^0.5 is used as weight in output_metric}

\item{DB.OUTPUT}{A logical variable: with value "T", output database is output; "F" is required in solver}
}
\value{
goodness-of-fit (GoF)
}
\description{
Generate the FUN (returning GoF) to be used in hindcast optimization
}
